Added relations between database columns and tables

+++ modules/Source/Database/Migrations/2024_01_13_140928_create_sources_table.php
@@ -15,10 +15,10 @@ return new class extends Migration {
             $table->string('slug', 255)->unique();
             $table->string('description', 5000)->nullable();
             $table->string('url', 255);
-            $table->tinyInteger('provider_id');
-            $table->integer('category_id')->nullable();
-            $table->integer('country_id')->nullable();
-            $table->integer('language_id')->nullable();
+            $table->unsignedBigInteger('provider_id');
+            $table->unsignedBigInteger('category_id')->nullable();
+            $table->unsignedBigInteger('country_id')->nullable();
+            $table->unsignedBigInteger('language_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
